Database manager added
