+++ environment.yml
@@ -9,6 +9,7 @@ dependencies:
   - numpy
   - scipy
   - pandas
+  - pandas-stubs
   # building and speed
   - build
   - setuptools
@@ -39,6 +40,7 @@ dependencies:
   # Data handling
   - sqlite
   - sqlalchemy
+  - sqlalchemy2-stubs
   - hdf5
   - h5py
   - netcdf4
