--- requirements-dev.txt
@@ -1,7 +1,0 @@
-# linting
-black
-flake8
-pylint
-# Testing
-pytest
-pytest-cov
